% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataMatching.R
\name{predict10}
\alias{predict10}
\title{Prediction with Top Ten Features}
\usage{
predict10(X, b)
}
\arguments{
\item{X}{}

\item{b}{a vector of coefficeints obtained from the \code{topten} function}
}
\value{
a numeric vector contatin the predictec values
}
\description{
Prediction with Top Ten Features
}
\details{
This function takes a set of coefficients produced by the \code{topten}
function and makes a prediction for each ot the values
}

