% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataMatching.R
\name{topten}
\alias{topten}
\title{Building a Model with Top ten Features}
\usage{
topten(x, y)
}
\arguments{
\item{x}{a n X p matric of n observations and p predictors}

\item{y}{a vector of length n representing the response}
}
\value{
a vector of coefficients from the final fitted moel with top 10 features
}
\description{
This function develops a prediction algorithm based on the top 10 features
in 'x' that are most predictive of 'y'
}
\details{
This function runs a univariate regression
}
\author{
Daniel Emaasit
}
\seealso{
\code{lm}
}

